+++ cranelift/filetests/src/lib.rs
@@ -57,7 +57,6 @@ mod test_shrink;
 mod test_simple_gvn;
 mod test_simple_preopt;
 mod test_unwind;
-mod test_vcode;
 mod test_verifier;
 
 /// The result of running the test in a file.
@@ -141,7 +140,6 @@ fn new_subtest(parsed: &TestCommand) -> subtest::SubtestResult<Box<dyn subtest::
         "simple-gvn" => test_simple_gvn::subtest(parsed),
         "simple_preopt" => test_simple_preopt::subtest(parsed),
         "unwind" => test_unwind::subtest(parsed),
-        "vcode" => test_vcode::subtest(parsed),
         "verifier" => test_verifier::subtest(parsed),
         _ => Err(format!("unknown test command '{}'", parsed.command)),
     }